// packages/service-api/sdk/model/apiChatCompletionPost200Response.ts
/**
 * Test swagger
 * testing the fastify swagger api
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ApiChatCompletionPost200ResponseChoicesInner } from './apiChatCompletionPost200ResponseChoicesInner';
import { ApiChatCompletionPost200ResponseUsage } from './apiChatCompletionPost200ResponseUsage';

export class ApiChatCompletionPost200Response {
    'id': string;
    'object': ApiChatCompletionPost200Response.ObjectEnum;
    'created': number;
    'model': string;
    'choices': Array<ApiChatCompletionPost200ResponseChoicesInner>;
    'usage': ApiChatCompletionPost200ResponseUsage;
    'systemFingerprint': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "object",
            "baseName": "object",
            "type": "ApiChatCompletionPost200Response.ObjectEnum"
        },
        {
            "name": "created",
            "baseName": "created",
            "type": "number"
        },
        {
            "name": "model",
            "baseName": "model",
            "type": "string"
        },
        {
            "name": "choices",
            "baseName": "choices",
            "type": "Array<ApiChatCompletionPost200ResponseChoicesInner>"
        },
        {
            "name": "usage",
            "baseName": "usage",
            "type": "ApiChatCompletionPost200ResponseUsage"
        },
        {
            "name": "systemFingerprint",
            "baseName": "system_fingerprint",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ApiChatCompletionPost200Response.attributeTypeMap;
    }
}

export namespace ApiChatCompletionPost200Response {
    export enum ObjectEnum {
        ChatCompletion = <any> 'chat.completion'
    }
}