// packages/service-api/sdk/model/apiChatCompletionPost200ResponseChoicesInner.ts
/**
 * Test swagger
 * testing the fastify swagger api
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ApiChatCompletionPost200ResponseChoicesInnerMessage } from './apiChatCompletionPost200ResponseChoicesInnerMessage';

export class ApiChatCompletionPost200ResponseChoicesInner {
    'index': number;
    'message': ApiChatCompletionPost200ResponseChoicesInnerMessage;
    'logprobs': object | null;
    'finishReason': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "index",
            "baseName": "index",
            "type": "number"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "ApiChatCompletionPost200ResponseChoicesInnerMessage"
        },
        {
            "name": "logprobs",
            "baseName": "logprobs",
            "type": "object"
        },
        {
            "name": "finishReason",
            "baseName": "finish_reason",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ApiChatCompletionPost200ResponseChoicesInner.attributeTypeMap;
    }
}