// packages/service-api/sdk/model/apiChatCompletionPost200ResponseChoicesInnerMessage.ts
/**
 * Test swagger
 * testing the fastify swagger api
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class ApiChatCompletionPost200ResponseChoicesInnerMessage {
    'role': string;
    'content': string;
    'refusal': string | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "role",
            "baseName": "role",
            "type": "string"
        },
        {
            "name": "content",
            "baseName": "content",
            "type": "string"
        },
        {
            "name": "refusal",
            "baseName": "refusal",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ApiChatCompletionPost200ResponseChoicesInnerMessage.attributeTypeMap;
    }
}