// packages/service-api/sdk/model/apiChatCompletionPost200ResponseUsage.ts
/**
 * Test swagger
 * testing the fastify swagger api
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ApiChatCompletionPost200ResponseUsageCompletionTokensDetails } from './apiChatCompletionPost200ResponseUsageCompletionTokensDetails';

export class ApiChatCompletionPost200ResponseUsage {
    'promptTokens': number;
    'completionTokens': number;
    'totalTokens': number;
    'completionTokensDetails': ApiChatCompletionPost200ResponseUsageCompletionTokensDetails;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "promptTokens",
            "baseName": "prompt_tokens",
            "type": "number"
        },
        {
            "name": "completionTokens",
            "baseName": "completion_tokens",
            "type": "number"
        },
        {
            "name": "totalTokens",
            "baseName": "total_tokens",
            "type": "number"
        },
        {
            "name": "completionTokensDetails",
            "baseName": "completion_tokens_details",
            "type": "ApiChatCompletionPost200ResponseUsageCompletionTokensDetails"
        }    ];

    static getAttributeTypeMap() {
        return ApiChatCompletionPost200ResponseUsage.attributeTypeMap;
    }
}