// packages/service-api/sdk/model/apiChatCompletionPost200ResponseUsageCompletionTokensDetails.ts
/**
 * Test swagger
 * testing the fastify swagger api
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class ApiChatCompletionPost200ResponseUsageCompletionTokensDetails {
    'reasoningTokens': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "reasoningTokens",
            "baseName": "reasoning_tokens",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return ApiChatCompletionPost200ResponseUsageCompletionTokensDetails.attributeTypeMap;
    }
}