// packages/service-api/sdk/model/apiChatCompletionPostRequest.ts
/**
 * Test swagger
 * testing the fastify swagger api
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class ApiChatCompletionPostRequest {
    'prompt': string;
    'model'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "prompt",
            "baseName": "prompt",
            "type": "string"
        },
        {
            "name": "model",
            "baseName": "model",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ApiChatCompletionPostRequest.attributeTypeMap;
    }
}