// packages/service-api/sdk/model/def3.ts
/**
 * Test swagger
 * testing the fastify swagger api
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* The service could not serve the request because of a client side related error
*/
export class Def3 {
    'error': Def3.ErrorEnum;
    'statusCode': Def3.StatusCodeEnum;
    'message': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "error",
            "baseName": "error",
            "type": "Def3.ErrorEnum"
        },
        {
            "name": "statusCode",
            "baseName": "statusCode",
            "type": "Def3.StatusCodeEnum"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return Def3.attributeTypeMap;
    }
}

export namespace Def3 {
    export enum ErrorEnum {
        BadRequest = <any> 'Bad Request'
    }
    export enum StatusCodeEnum {
        NUMBER_400 = <any> 400
    }
}