// packages/service-api/sdk/model/def4.ts
/**
 * Test swagger
 * testing the fastify swagger api
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* Response for not authenticated users
*/
export class Def4 {
    'error': Def4.ErrorEnum;
    'message': Def4.MessageEnum;
    'statusCode': Def4.StatusCodeEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "error",
            "baseName": "error",
            "type": "Def4.ErrorEnum"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "Def4.MessageEnum"
        },
        {
            "name": "statusCode",
            "baseName": "statusCode",
            "type": "Def4.StatusCodeEnum"
        }    ];

    static getAttributeTypeMap() {
        return Def4.attributeTypeMap;
    }
}

export namespace Def4 {
    export enum ErrorEnum {
        Unauthorized = <any> 'Unauthorized'
    }
    export enum MessageEnum {
        NotValid = <any> 'auth_header_not_valid',
        Missed = <any> 'auth_header_missed'
    }
    export enum StatusCodeEnum {
        NUMBER_401 = <any> 401
    }
}